# docs/OpenAPI.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Event Ticketing & Seat Reservation API",
    "description": "Backend API for discovering events, reserving seats, and managing concurrent bookings with JWT authentication and role-based access.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/admin/events/{eventId}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get any event details",
        "description": "Retrieves the details of any event system-wide.",
        "operationId": "adminGetEventDetails",
        "parameters": [
          { "name": "eventId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Event details retrieved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventDetailDTO" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Admin"],
        "summary": "Update any event",
        "description": "Force updates any event system-wide, bypassing ownership checks.",
        "operationId": "adminUpdateEvent",
        "parameters": [
          { "name": "eventId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateEventRequestDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventDetailDTO" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete any event",
        "description": "Force deletes an event and cascades to its seats and reservations.",
        "operationId": "adminDeleteEvent",
        "parameters": [
          { "name": "eventId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Event deleted successfully" }
        }
      }
    },
    "/api/admin/events/{eventId}/reservations": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get reservations by event",
        "description": "Retrieves all reservations for a specific event system-wide.",
        "operationId": "adminGetReservationsByEvent",
        "parameters": [
          { "name": "eventId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Reservations retrieved",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ReservationSummaryDTO" } }
              }
            }
          }
        }
      }
    },
    "/api/admin/venues": {
      "post": {
        "tags": ["Admin"],
        "summary": "Create a new venue",
        "description": "Adds a new venue to the platform that organizers can use to host events.",
        "operationId": "adminCreateVenue",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateVenueDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Venue created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VenueSummaryDTO" }
              }
            }
          }
        }
      }
    },
    "/api/admin/venues/{venueId}": {
      "put": {
        "tags": ["Admin"],
        "summary": "Update a venue",
        "description": "Updates details like location or stars for a specific venue.",
        "operationId": "adminUpdateVenue",
        "parameters": [
          { "name": "venueId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateVenueDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Venue updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VenueSummaryDTO" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete a venue",
        "description": "Removes a venue from the system.",
        "operationId": "adminDeleteVenue",
        "parameters": [
          { "name": "venueId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Venue deleted successfully" }
        }
      }
    },
    "/api/admin/venues/{venueId}/reservations": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get reservations by venue",
        "description": "Retrieves all reservations tied to events hosted at a specific venue.",
        "operationId": "adminGetReservationsByVenue",
        "parameters": [
          { "name": "venueId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Reservations retrieved",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ReservationSummaryDTO" } }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/reservations/{reservationId}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get any reservation details",
        "description": "Retrieves detailed information for any specific reservation system-wide.",
        "operationId": "adminGetReservationById",
        "parameters": [
          { "name": "reservationId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Reservation retrieved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReservationDetailDTO" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete or force-cancel reservation",
        "description": "Forces the cancellation and deletion of any reservation system-wide.",
        "operationId": "adminCancelReservation",
        "parameters": [
          { "name": "reservationId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Reservation cancelled/deleted" }
        }
      }
    },
...
    "/api/events/{eventId}": {
      "get": {
        "tags": ["Public Events"],
        "summary": "Get event details",
        "description": "Retrieves the full public details of a specific event.",
        "operationId": "publicGetEventById",
        "parameters": [
          { "name": "eventId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Event details retrieved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventDetailDTO" }
              }
            }
          }
        }
      }
    },
...
    "/api/venues/{venueId}": {
      "get": {
        "tags": ["Public Venues"],
        "summary": "Get venue details",
        "description": "Retrieves the public details of a specific venue.",
        "operationId": "publicGetVenueById",
        "parameters": [
          { "name": "venueId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Venue details retrieved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VenueSummaryDTO" }
              }
            }
          }
        }
      }
    },
    "/api/venues/{venueId}/events": {
      "get": {
        "tags": ["Public Venues"],
        "summary": "List events for a venue",
        "description": "Retrieves all events scheduled to take place at a specific venue.",
        "operationId": "publicGetEventsByVenueId",
        "parameters": [
          { "name": "venueId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Events retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/EventSummaryDTO" } }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEventRequestDTO": {
        "type": "object",
        "properties": {
          "venueId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "theme": { "type": "string" },
          "description": { "type": "string" },
          "startTimestamp": { "type": "string", "format": "date-time" },
          "endTimestamp": { "type": "string", "format": "date-time" },
          "ticketPrice": { "type": "number" },
          "seatCapacity": { "type": "integer", "format": "int64" }
        },
        "example": {
          "venueId": "550e8400-e29b-41d4-a716-446655440000",
          "name": "Spring Boot Global Summit 2026",
          "theme": "Technology & Backend",
          "description": "A comprehensive deep dive into microservices and scaling with Spring Boot 3.",
          "startTimestamp": "2026-11-15T09:00:00Z",
          "endTimestamp": "2026-11-15T18:00:00Z",
          "ticketPrice": 150.00,
          "seatCapacity": 200
        }
      },
      "EventDetailDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "theme": { "type": "string" },
          "venueId": { "type": "string", "format": "uuid" },
          "venueName": { "type": "string" },
          "venueStars": { "type": "integer", "format": "int32" },
          "seatCapacity": { "type": "integer", "format": "int32" },
          "ticketPrice": { "type": "number" },
          "startTimestamp": { "type": "string", "format": "date-time" },
          "organizerName": { "type": "string" },
          "endTimestamp": { "type": "string", "format": "date-time" },
          "description": { "type": "string" }
        },
        "example": {
          "id": "e44d32f1-9b1a-4c22-8176-a19e34e56b82",
          "name": "Spring Boot Global Summit 2026",
          "theme": "Technology & Backend",
          "venueId": "550e8400-e29b-41d4-a716-446655440000",
          "venueName": "Grand Tech Arena",
          "venueStars": 5,
          "seatCapacity": 200,
          "ticketPrice": 150.00,
          "startTimestamp": "2026-11-15T09:00:00Z",
          "endTimestamp": "2026-11-15T18:00:00Z",
          "organizerName": "Tech Conferences LLC",
          "description": "A comprehensive deep dive into microservices and scaling with Spring Boot 3."
        }
      },
      "CreateVenueDTO": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "location": { "type": "string" },
          "stars": { "type": "integer", "format": "int32" },
          "description": { "type": "string" }
        },
        "example": {
          "name": "Grand Tech Arena",
          "location": "Downtown Tech District, CA",
          "stars": 5,
          "description": "A state-of-the-art auditorium designed for large-scale technology summits."
        }
      },
      "VenueSummaryDTO": {
        "type": "object",
        "properties": {
          "venueId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "location": { "type": "string" },
          "stars": { "type": "integer", "format": "int32" },
          "description": { "type": "string" }
        },
        "example": {
          "venueId": "550e8400-e29b-41d4-a716-446655440000",
          "name": "Grand Tech Arena",
          "location": "Downtown Tech District, CA",
          "stars": 5,
          "description": "A state-of-the-art auditorium designed for large-scale technology summits."
        }
      },
      "EventSummaryDTO": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "theme": { "type": "string" },
          "venueId": { "type": "string", "format": "uuid" },
          "venueName": { "type": "string" },
          "venueStars": { "type": "integer", "format": "int32" },
          "seatCapacity": { "type": "integer", "format": "int32" },
          "ticketPrice": { "type": "number" },
          "startTimestamp": { "type": "string", "format": "date-time" }
        },
        "example": {
          "id": "e44d32f1-9b1a-4c22-8176-a19e34e56b82",
          "name": "Spring Boot Global Summit 2026",
          "theme": "Technology & Backend",
          "venueId": "550e8400-e29b-41d4-a716-446655440000",
          "venueName": "Grand Tech Arena",
          "venueStars": 5,
          "seatCapacity": 200,
          "ticketPrice": 150.00,
          "startTimestamp": "2026-11-15T09:00:00Z"
        }
      },
...
      "ReservationDetailDTO": {
        "type": "object",
        "properties": {
          "reservationId": { "type": "string", "format": "uuid" },
          "eventId": { "type": "string", "format": "uuid" },
          "eventName": { "type": "string" },
          "eventStartTime": { "type": "string", "format": "date-time" },
          "eventEndTime": { "type": "string", "format": "date-time" },
          "venueName": { "type": "string" },
          "seatsNumbers": { "type": "array", "items": { "type": "string" } },
          "totalPrice": { "type": "number" },
          "status": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "example": {
          "reservationId": "9b1c72f4-a83b-4b11-91a5-813c9e651d2e",
          "eventId": "e44d32f1-9b1a-4c22-8176-a19e34e56b82",
          "eventName": "Spring Boot Global Summit 2026",
          "eventStartTime": "2026-11-15T09:00:00Z",
          "eventEndTime": "2026-11-15T18:00:00Z",
          "venueName": "Grand Tech Arena",
          "seatsNumbers": ["A1", "A2"],
          "totalPrice": 300.00,
          "status": "CONFIRMED",
          "createdAt": "2026-10-01T14:32:00Z"
        }
      },
      "ReservationSummaryDTO": {
        "type": "object",
        "properties": {
          "reservationId": { "type": "string", "format": "uuid" },
          "eventId": { "type": "string", "format": "uuid" },
          "eventName": { "type": "string" },
          "status": { "type": "string" },
          "totalPrice": { "type": "number", "format": "double" }
        },
        "example": {
          "reservationId": "9b1c72f4-a83b-4b11-91a5-813c9e651d2e",
          "eventId": "e44d32f1-9b1a-4c22-8176-a19e34e56b82",
          "eventName": "Spring Boot Global Summit 2026",
          "status": "CONFIRMED",
          "totalPrice": 300.00
        }
      },
...
    }
  }
}